refactor(app): clean up deck config dispatching
