--- app/src/organisms/LabwarePositionCheck/LPCFlows/hooks/useUpdateLPCStore/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './getActivePipetteId'
